Refactor FakeData thread execution
