+++ .vscode/launch.json
@@ -18,7 +18,7 @@
             "name": "Create FakeData <jobwait>",
             "request": "launch",
             "cwd": "${workspaceFolder}",
-            "mainClass": "com.jobwait.fakedata.FakeData",
+            "mainClass": "com.jobwait.fakedata.FakeDataExecutor",
             "projectName": "jobwait",
             "args": "createFakeData",
             "envFile": "${workspaceFolder}/.env",
